Add files via upload
In this submission, my intent was to use JavaScript library (papaparse) and HTML to read the excel dataset which I had converted to a .csv file. However, when I uploaded the dataset so as to view the data in JSON format in the console, there was no output and I could not really establish what I was not getting right.  With familiarity in JavaScript libraries and a bit of guidance, I understand am capable of delivering the expected requirements in this challenge.